Current img's index, VPrintHud
NEW: now maae keeps track of the current img's index, and shows it to us.
TODO: fix index when creating a new img

NEW: now you can call PrintHud with formatted strings, like printf: just
use the VPrintHud function! Had to change the arguments order because of
that, and to be the same, changed it in the PrintHud aswell